Remove reference to `aws ec2 create-vpc`
When creating an AWS BYOVPC cluster the customer typically has 1 of 2
scenarios:
1) They already have a VPC created that they want to use
2) They are using our terraform code to create a VPC following our specifications

Never would they want to create the VPC using an `aws ec2 create-vpc` command as
specified here. The command even includes a hard-coded profile name that is only
valid internal to Redpanda. We should not expose this profile name outside
of Redpanda.